fix compile error on darwin
